+++ tests/test-ci/test-httpd.R
@@ -80,7 +80,10 @@ http_request_full = function(host, port, method, path, body = NULL, extra_header
 
 url_port = function(url) as.integer(sub('^https?://[^:]+:([0-9]+)/.*$', '\\1', url))
 
-if (!is.null(port <- random_port(error = FALSE))) {
+if (
+  has_fun('serverSocket') && has_fun('socketAccept') &&
+  !is.null(port <- random_port(error = FALSE))
+) {
 
   # Run the proxy-backed app in a separate R process so requests are sent from
   # this R session to reduce random timing issues in CI.